Move rule_src_parser inside cpp_rule_preprocessor
